+++ ViewModels/SingleVolunteerViewModel.cs
@@ -6,7 +6,7 @@ namespace BenLearns.ViewModels
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Role { get; set; }
-        public int RoleID { get; set; }
+        public long RoleID { get; set; }
 
 
     }
